+++ sdk/communication/azure-communication-callautomation/azure/communication/callautomation/_generated/aio/operations/__init__.py
@@ -9,7 +9,6 @@
 from ._operations import AzureCommunicationCallAutomationServiceOperationsMixin
 from ._operations import CallConnectionOperations
 from ._operations import CallMediaOperations
-from ._operations import CallDialogOperations
 from ._operations import CallRecordingOperations
 
 from ._patch import __all__ as _patch_all
@@ -20,8 +19,7 @@ __all__ = [
     "AzureCommunicationCallAutomationServiceOperationsMixin",
     "CallConnectionOperations",
     "CallMediaOperations",
-    "CallDialogOperations",
     "CallRecordingOperations",
 ]
-__all__.extend([p for p in _patch_all if p not in __all__])
+__all__.extend([p for p in _patch_all if p not in __all__])  # pyright: ignore
 _patch_sdk()