+++ sdk/communication/azure-communication-callautomation/tests/test_e2e_callautomation_client.py
@@ -5,6 +5,7 @@
 # --------------------------------------------------------------------------
 from datetime import timedelta
 import time
+import pytest
 
 from devtools_testutils import recorded_by_proxy
 
@@ -61,6 +62,41 @@ class TestCallAutomationClientAutomatedLiveTest(CallAutomationRecordedTestCase):
 
         if cancel_add_participant_succeeded_event is None:
             raise ValueError("Caller CancelAddParticipantSucceeded event is None")
+
+        self.terminate_call(unique_id)
+        return
+    
+    @pytest.mark.skip(reason="""Playback fails for same event type triggered and test recording code
+                       takes the event type has the dictionary it fails to recording call connected event for the connect api""")
+    @recorded_by_proxy
+    def test_create_VOIP_call_and_connect_call_then_hangup(self):
+        # try to establish the call
+        caller = self.identity_client.create_user()
+        target = self.identity_client.create_user()
+        unique_id, call_connection, _, call_automation_client, callback_url = self.establish_callconnection_voip_connect_call(caller, target)
+
+        # check returned events
+        connected_event = self.check_for_event('CallConnected', call_connection._call_connection_id, timedelta(seconds=15))
+        participant_updated_event = self.check_for_event('ParticipantsUpdated', call_connection._call_connection_id, timedelta(seconds=15))
+
+        if connected_event is None:
+            raise ValueError("Caller CallConnected event is None")
+        if participant_updated_event is None:
+            raise ValueError("Caller ParticipantsUpdated event is None")
+
+        call_connection_properties = call_connection.get_call_properties()
+        server_call_id = call_connection_properties.server_call_id
+
+        # connect call request.
+        connect_call_connection = call_automation_client.connect_call(
+            server_call_id=server_call_id,
+            callback_url=callback_url,
+            )
+
+        # check returned call connected events
+        connect_call_connected_event = self.check_for_event('CallConnected', connect_call_connection.call_connection_id, timedelta(seconds=15))
+        if connect_call_connected_event is None:
+            raise ValueError("Caller CallConnected event is None")
 
         self.terminate_call(unique_id)
         return