+++ sdk/communication/azure-communication-callautomation/CHANGELOG.md
@@ -10,6 +10,20 @@
 
 ### Other Changes
 
+## 1.3.0b1 (2024-08-02)
+
+### Features Added
+
+- Support multiple play sources for Play and Recognize
+- Support for PlayStarted event in Play/Recognize
+- Support for the real time transcription
+- Monetization for real-time transcription and audio streaming
+- Hold and Unhold the participant
+- Support to manage the rooms/servercall/group call using connect API
+- Support for the audio streaming
+- Expose original PSTN number target from incoming call event in call connection properties
+- Support for VoIP to PSTN transfer scenario
+
 ## 1.2.0 (2024-04-15)
 
 ### Features Added